Create choice buttons and event listeners
In order to create a UI for the rock, paper, scissors game, there
needs to be three buttons each with one of the options, and when
one of the buttons is clicked a round of the game is played. These
buttons have been created in index.html and event listeners have
been attached to each button. When a button is pressed, the
playRound function is called with the user's choice and a random
computer choice generated by the getComputerChoice function in
script.js. The result of the round is printed in the console.