Update UI to show result of each round
It would be useful to the player to know the results of each
round of the game. A div element has been added to index.html
which is updated in the event listener of each button after
each round of the game to show whether the player won or lost
the round.