Update player and computer score in UI
It would be useful for the player to know their own running score
and the computer's running score. Two div elements have been
added to index.html that display the player's running score and
the computer's running score. The text content of these divs are
updated after every round in the updateScore function, if needed.