Announce winner at end of game
Before, an infinite number of rounds could be played. There was
no end to the game. Now, a count is kept of the number of rounds
played. After five rounds are played, the game ends, the option
buttons can no longer be clicked, and a message is displayed that
states if the player won or lost the game.