Add Reset Game button
Before, the game ended after five rounds, but the only way to
reset the game was to refresh the page. Now, there is a Reset
Game button. It is only clickable after all five rounds of the
game are played. When it is clicked, the player score, computer
score, and number of rounds are reset, the option buttons are
once again clickable, and the game over message from the previous
game is removed.